+++ GrammarNazi.Core/Clients/MeaningCloudSentimentAnalysisApiClient.cs
@@ -1,6 +1,6 @@
 ﻿using GrammarNazi.Domain.Clients;
-using GrammarNazi.Domain.Entities.LanguageIdentificationAPI;
-using GrammarNazi.Domain.Entities.LanguageIdentificationAPI.SentimentAnalysis;
+using GrammarNazi.Domain.Entities.MeaningCloudAPI;
+using GrammarNazi.Domain.Entities.MeaningCloudAPI.SentimentAnalysis;
 using GrammarNazi.Domain.Entities.Settings;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
