+++ GrammarNazi.Core/Clients/MeganingCloudLangApiClient.cs
@@ -1,5 +1,5 @@
 ﻿using GrammarNazi.Domain.Clients;
-using GrammarNazi.Domain.Entities.LanguageIdentificationAPI;
+using GrammarNazi.Domain.Entities.MeaningCloudAPI;
 using GrammarNazi.Domain.Entities.Settings;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
