+++ GrammarNazi.Domain/Entities/MeaningCloudAPI/SentimentAnalysis/SentimentAnalysisResult.cs
@@ -1,8 +1,8 @@
-﻿using GrammarNazi.Domain.Entities.LanguageIdentificationAPI.SentimentAnalysis;
+﻿using GrammarNazi.Domain.Entities.MeaningCloudAPI.SentimentAnalysis;
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace GrammarNazi.Domain.Entities.LanguageIdentificationAPI
+namespace GrammarNazi.Domain.Entities.MeaningCloudAPI
 {
     public class SentimentAnalysisResult
     {
